+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ENHANCEMENTS:
 
-* Supports `dynamic` block for `tags` and `labels`
+* Supports `dynamic` block for `tags`, `labels` and `regions_config`
 
 ## 1.0.0 (Mar 6, 2025)