+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## (Unreleased)
 
+## 1.1.0 (Mar 25, 2025)
+
 ENHANCEMENTS:
 
 * Supports `dynamic` blocks for `tags`, `labels`, `regions_config` and `replication_specs`